+++ source/features/update-pr-from-base-branch.tsx
@@ -1,6 +1,6 @@
 import React from 'dom-chef';
 import select from 'select-dom';
-import splitOnFirst from 'split-on-first';
+
 import {AlertIcon} from '@primer/octicons-react';
 import * as pageDetect from 'github-url-detection';
 import delegate, {DelegateEvent} from 'delegate-it';
@@ -8,42 +8,12 @@ import delegate, {DelegateEvent} from 'delegate-it';
 import features from '../feature-manager';
 import observe from '../helpers/selector-observer';
 import * as api from '../github-helpers/api';
+import {getBranches} from '../github-helpers/pr-branches';
 import getPrInfo from '../github-helpers/get-pr-info';
 import {getConversationNumber} from '../github-helpers';
 
 const selectorForPushablePRNotice = '.merge-pr > .color-fg-muted:first-child';
 
-type PrReference = {
-	/** @example fregante/mem:main */
-	full: string;
-
-	/** @example "main" on same-repo PRs, "fregante:main" on cross-repo PRs  */
-	local: string;
-
-	/** @example fregante */
-	owner: string;
-
-	/** @example mem */
-	name: string;
-
-	/** @example main */
-	branch: string;
-};
-
-function parseReference(referenceElement: HTMLElement): PrReference {
-	const {title: full, textContent: local} = referenceElement;
-	const [nameWithOwner, branch] = splitOnFirst(full, ':') as [string, string];
-	const [owner, name] = nameWithOwner.split(':');
-	return {full, owner, name, branch, local: local!.trim()};
-}
-
-// TODO: Use in more places, like anywhere '.base-ref' appears
-export function getBranches(): {base: PrReference; head: PrReference} {
-	return {
-		base: parseReference(select('.base-ref')!),
-		head: parseReference(select('.head-ref')!),
-	};
-}
 
 async function mergeBranches(): Promise<AnyObject> {
 	return api.v3(`pulls/${getConversationNumber()!}/update-branch`, {
